requirements.txt
# Core Dependencies
playwright==1.30.0
beautifulsoup4==4.9.3
langchain==0.0.300
openai==1.0.0
fastapi==0.68.0
uvicorn==0.15.0 #any uvicorn error - pip install fastapi uvicorn
streamlit==1.2.0
aiohttp==3.8.0
pydantic==2.0.0
redis==4.0.0
loguru==0.5.3
python-dotenv==1.0.0

# Testing Dependencies
pytest==7.0.0
pytest-asyncio==0.20.0
pytest-cov==4.0.0

# Development Dependencies
black==22.0.0
isort==5.0.0
mypy==1.0.0
pre-commit==2.17.0